file didn't commit originally

+++ src/pages/api/listings/[id].ts
@@ -3,6 +3,7 @@ import prisma from "lib/db";
 import ListingsDataTable from "@/models/listings";
 import { listings } from "@prisma/client";
 import listingDataTable from "lib/listingInstance";
+import { getSession } from "next-auth/react";
 
 type Message = {
   message: string;
@@ -49,6 +50,7 @@ async function updateListing(
   res: NextApiResponse<Message>
 ) {
   try {
+    await authorize(req);
     await listingDataTable.updateListing(req.body);
   } catch (error) {
     return res.status(403).send({ message: String(error) });
@@ -62,10 +64,40 @@ async function deleteListing(
   res: NextApiResponse<Message>
 ) {
   try {
+    await authorize(req);
     await listingDataTable.deleteListing(req.body);
   } catch (error) {
     return res.status(403).send({ message: String(error) });
   }
 
   return res.status(200).send({ message: "deleted listing" });
+}
+
+async function authorize(req: NextApiRequest) {
+  const session = await getSession({ req });
+
+  if (!session) {
+    throw new Error("You must be logged in to perform this action.");
+  }
+
+  if (!session.user || !session.user.email) {
+    throw new Error("User email not found in session.");
+  }
+
+  const user = await prisma.users.findUnique({
+    where: { email: session.user.email },
+  });
+
+  if (!user || !user.verified) {
+    throw new Error("You must be a verified user to perform this action.");
+  }
+
+  const listingId = req.query.id;
+  const listing = await prisma.listings.findUnique({
+    where: { listing_id: Number(listingId) },
+  });
+
+  if (!listing || listing.host_id !== user.user_id) {
+    throw new Error("You do not have permission to perform this action.");
+  }
 }
